seata 的学习 时间

+++ cloud-alibaba-seata-account-2003/src/main/java/com/chen/cloud/alibaba/dao/AccountDao.java
@@ -2,13 +2,11 @@ package com.chen.cloud.alibaba.dao;
 
 import org.apache.ibatis.annotations.Param;
 
-import java.math.BigDecimal;
-
 /**
  * @author chenpc
  * @version 1.0
  * @since 2021/6/8/06/08  17:09
  */
 public interface AccountDao {
-    void decrease(@Param("userId") Long userId, @Param("money") BigDecimal money);
+    void decrease(@Param("userId") Long userId, @Param("money") Long money);
 }
