+++ cloud-alibaba-seata-account-2003/src/main/java/com/chen/cloud/alibaba/service/IAccountService.java
@@ -1,12 +1,10 @@
 package com.chen.cloud.alibaba.service;
 
-import java.math.BigDecimal;
-
 /**
  * @author chenpc
  * @version 1.0
  * @since 2021/6/8/06/08  17:13
  */
 public interface IAccountService {
-    void decrease(Long userId, BigDecimal money);
+    void decrease(Long userId, Long money);
 }
