+++ cloud-alibaba-seata-order-2001/src/main/java/com/chen/cloud/alibaba/service/impl/AccountServiceImpl.java
@@ -3,7 +3,6 @@ package com.chen.cloud.alibaba.service.impl;
 import com.chen.cloud.alibaba.bean.CommonResult;
 import com.chen.cloud.alibaba.service.IAccountService;
 import org.springframework.stereotype.Component;
-import java.math.BigDecimal;
 
 /**
  * @author chenpc
@@ -13,7 +12,7 @@ import java.math.BigDecimal;
 @Component
 public class AccountServiceImpl implements IAccountService {
     @Override
-    public CommonResult decrease(Long userId, BigDecimal money) {
+    public CommonResult decrease(Long userId, Long money) {
         return new CommonResult(500, "账户服务调用失败-->请重试");
     }
 }
